Extract editor main front matter manager
