+++ assets/press-system.json
@@ -1,13 +1,13 @@
 {
   "schemaVersion": 1,
   "type": "press-system",
-  "version": "3.4.86",
-  "tag": "v3.4.86",
+  "version": "3.4.87",
+  "tag": "v3.4.87",
   "upgradeFrom": {
     "ranges": [
-      ">=3.4.85 <3.4.86"
+      ">=3.4.86 <3.4.87"
     ],
     "allowUnknownSource": false,
-    "message": "Update to v3.4.85 first."
+    "message": "Update to v3.4.86 first."
   }
 }
