+++ scripts/test-system-release-package.sh
@@ -288,6 +288,11 @@ if ! grep -qx "press-system-${version}/assets/js/editor-main-frontmatter-label-w
   exit 1
 fi
 
+if ! grep -qx "press-system-${version}/assets/js/editor-main-frontmatter-manager.js" "${entries_file}"; then
+  echo "expected package to include editor main article front matter manager code" >&2
+  exit 1
+fi
+
 if ! grep -qx "press-system-${version}/assets/js/editor-main-tabs-metadata-manager.js" "${entries_file}"; then
   echo "expected package to include editor main tabs metadata manager code" >&2
   exit 1
